requirements.txt
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.14.0
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audeer==2.0.0
audiofile==1.4.0
audioread==3.0.1
audmath==1.4.0
audtorch==0.6.4
av==12.3.0
backports.tarfile==1.2.0
boto3==1.34.113
botocore==1.34.113
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1648028898966/work
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
colorlog==6.9.0
contourpy==1.2.1
crcmod==1.7
cryptography==43.0.0
cycler==0.12.1
daal==2024.6.0
daal4py==2024.6.0
datasets==2.16.0
decorator==4.4.2
dill==0.3.6
docutils==0.21.2
einops==0.7.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.1.0
filelock==3.9.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2023.10.0
gast==0.6.0
gensim==4.3.2
greenlet==3.1.1
h5py==3.10.0
huggingface-hub==0.23.2
idna==3.6
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_metadata==8.2.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipython==8.18.1
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.2
jeepney==0.8.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
keyring==25.2.1
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl-fft==1.3.0
mkl-random==1.0.2
mkl-service==2.3.0
mmsdk @ file:///home/liu70kg/PycharmProjects/CMU-MultimodalSDK-main
modelscope==1.18.1
more-itertools==10.3.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.14
natsort==8.4.0
networkx==3.2.1
nh3==0.2.18
numba==0.59.1
numexpr @ file:///tmp/build/80754af9/numexpr_1614798502122/work
numpy==1.22.0
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
opencc-python-reimplemented==0.1.7
openpyxl==3.1.5
optuna==4.1.0
oss2==2.19.0
packaging @ file:///croot/packaging_1710807400464/work
pandas==1.4.2
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
pkginfo==1.10.0
platformdirs==4.2.0
plotly==5.24.1
pluggy==1.5.0
pooch==1.8.1
proglog==0.1.10
prompt_toolkit==3.0.48
protobuf==5.27.2
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.17
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
pytest==8.2.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///croot/pytz_1695131579487/work
PyYAML==6.0.1
readme_renderer==44.0
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
resampy==0.4.3
responses==0.18.0
rfc3986==2.0.0
rich==13.7.1
s3transfer==0.10.1
sacremoses==0.1.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
scikit-learn-intelex==2024.6.0
scipy==1.11.4
seaborn==0.13.2
SecretStorage==3.3.3
sentencepiece==0.2.0
shap==0.46.0
simplejson==3.19.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
slicer==0.0.8
smart-open==7.0.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
SQLAlchemy==2.0.36
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tbb==2021.13.1
tenacity==9.0.0
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
tiktoken==0.8.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.2.1+cu118
torch_geometric==2.5.3
torchaudio==2.2.1+cu118
torchinfo==1.8.0
torchprofile==0.0.4
torchvision==0.17.1+cu118
towhee==1.1.1
towhee.models==1.1.3
tqdm==4.66.2
traitlets==5.14.3
transformers==4.41.1
triton==2.2.0
twine==5.1.1
typing_extensions==4.8.0
urllib3==1.26.18
validators==0.23.1
wcwidth==0.2.13
wrapt==1.16.0
xxhash==3.5.0
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1
